Refactor core init
