+++ assets/core/index.js
@@ -7,12 +7,13 @@ import 'whatwg-fetch';
 import 'picturefill';
 
 import { timeOutAlerts } from 'common/utils/';
-import { initGoogleMaps, initialize } from './init.js';
+import { initGoogleMaps } from './maps';
+import init from './init';
 import './less/core.less';
 
 
 moment.locale('nb');
 
 initGoogleMaps();
-initialize();
+init();
 timeOutAlerts();
